setup/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
